added particle manager
